requirements.txt
django
django-cors-headers
djangorestframework
django-ical
django-taggit